Split `Query::class`.
